Fix leave lobby method

See merge request cse1105/2021-2022/team-repositories/oopp-group-60/repository-template!84

+++ commons/src/main/java/commons/entities/ActivityDTO.java
@@ -24,7 +24,7 @@ public class ActivityDTO implements DTO {
     /**
      * Energy consumption of the activity.
      */
-    protected int cost;
+    protected long cost;
 
     /**
      * URL of the picture of the activity.
